Add files via upload
Made modifications to the cs205_fina-2021.sh file. First I changed the file from a .sh to .awk to allow for the use of AWK. Second I added calculations to give the user the number of Pokemon, the average hp and average attack. Gave a Summary to the user via print when they run the file.